refactor!: move app lifecycle producer to demo app
