+++ ConfidenceDemoApp/src/main/java/com/example/confidencedemoapp/AndroidLifecycleEventProducer.kt
@@ -1,4 +1,4 @@
-package com.spotify.confidence
+package com.example.confidencedemoapp
 
 import android.app.Activity
 import android.app.Application
@@ -14,6 +14,9 @@ import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.ProcessLifecycleOwner
+import com.spotify.confidence.ConfidenceValue
+import com.spotify.confidence.Event
+import com.spotify.confidence.EventProducer
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
@@ -125,11 +128,6 @@ class AndroidLifecycleEventProducer(
         val currentVersion = ConfidenceValue.String(packageInfo?.versionName ?: "")
         val currentBuild = ConfidenceValue.String(packageInfo?.getVersionCode().toString() ?: "")
 
-        val addedContext = mapOf(
-            APP_VERSION_KEY to currentVersion,
-            APP_BUILD_KEY to currentBuild
-        )
-        updateContext(addedContext)
 
         val previousBuild: ConfidenceValue.String? = sharedPreferences
             .getString(APP_BUILD, null)
@@ -171,10 +169,6 @@ class AndroidLifecycleEventProducer(
         private const val APP_VERSION = "APP_VERSION"
         private const val APP_BUILD = "APP_BUILD"
         private const val LEGACY_APP_BUILD = "LEGACY_APP_BUILD"
-
-        // Context keys
-        private const val APP_VERSION_KEY = "app_version"
-        private const val APP_BUILD_KEY = "app_build"
 
         // Event keys
         private const val APP_INSTALLED_EVENT = "app-installed"
